requirements.txt
Flask
google-auth-oauthlib
google-auth
google-api-python-client
transformers
torch
nltk
python-dotenv
dateutil
